docs: convert BiaxialLineChart example to TypeScript (.tsx)
